+++ src/libtrx/game/gun/misc.c
@@ -251,3 +251,46 @@ void Gun_DrawFlash(LARA_GUN_TYPE weapon_type, CLIP clip)
         Object_DrawMesh(obj->mesh_idx, clip, false);
     }
 }
+
+void Gun_UpdateLaraMeshes(const GAME_OBJECT_ID obj_id)
+{
+#if TR_VERSION == 1
+    const bool lara_has_rifle = Inv_RequestItem(O_SHOTGUN_ITEM);
+#else
+    const bool lara_has_rifle = Inv_RequestItem(O_SHOTGUN_ITEM)
+        || Inv_RequestItem(O_HARPOON_ITEM) || Inv_RequestItem(O_M16_ITEM)
+        || Inv_RequestItem(O_GRENADE_ITEM);
+#endif
+    const bool lara_has_pistols = Inv_RequestItem(O_PISTOL_ITEM)
+        || Inv_RequestItem(O_MAGNUM_ITEM) || Inv_RequestItem(O_UZI_ITEM);
+
+    LARA_GUN_TYPE back_gun_type = LGT_UNARMED;
+    LARA_GUN_TYPE holsters_gun_type = LGT_UNARMED;
+
+    if (!lara_has_rifle && obj_id == O_SHOTGUN_ITEM) {
+        back_gun_type = LGT_SHOTGUN;
+#if TR_VERSION >= 2
+    } else if (!lara_has_rifle && obj_id == O_HARPOON_ITEM) {
+        back_gun_type = LGT_HARPOON;
+    } else if (!lara_has_rifle && obj_id == O_M16_ITEM) {
+        back_gun_type = LGT_M16;
+    } else if (!lara_has_rifle && obj_id == O_GRENADE_ITEM) {
+        back_gun_type = LGT_GRENADE;
+#endif
+    } else if (!lara_has_pistols && obj_id == O_PISTOL_ITEM) {
+        holsters_gun_type = LGT_PISTOLS;
+    } else if (!lara_has_pistols && obj_id == O_MAGNUM_ITEM) {
+        holsters_gun_type = LGT_MAGNUMS;
+    } else if (!lara_has_pistols && obj_id == O_UZI_ITEM) {
+        holsters_gun_type = LGT_UZIS;
+    }
+
+    if (back_gun_type != LGT_UNARMED) {
+        Gun_SetLaraBackMesh(back_gun_type);
+    }
+
+    if (holsters_gun_type != LGT_UNARMED) {
+        Gun_SetLaraHolsterLMesh(holsters_gun_type);
+        Gun_SetLaraHolsterRMesh(holsters_gun_type);
+    }
+}
