+++ src/libtrx/include/libtrx/game/gun/misc.h
@@ -1,5 +1,7 @@
 #pragma once
 
+// Private gun routines.
+
 #include "../items/types.h"
 #include "../lara/types.h"
 #include "../types.h"
@@ -14,6 +16,13 @@ typedef enum {
 void Gun_FindTargetPoint(const ITEM *item, GAME_VECTOR *target);
 void Gun_AimWeapon(const WEAPON_INFO *weapon, LARA_ARM *arm);
 void Gun_TargetInfo(const WEAPON_INFO *weapon);
+void Gun_UpdateLaraMeshes(GAME_OBJECT_ID obj_id);
+
+void Gun_GetNewTarget(const WEAPON_INFO *weapon);
+void Gun_ChangeTarget(const WEAPON_INFO *weapon);
+void Gun_HitTarget(ITEM *item, const GAME_VECTOR *hit_pos, int32_t damage);
+
+void Gun_DrawFlash(LARA_GUN_TYPE weapon_type, CLIP clip);
 
 #if TR_VERSION >= 2
 extern PROJECTILE_HIT Gun_SmashItems(
