--- src/tr1/game/gun/gun.c
@@ -1,35 +1,0 @@
-#include "game/gun.h"
-
-#include "game/inventory.h"
-
-#include <libtrx/config.h>
-#include <libtrx/game/random.h>
-
-void Gun_UpdateLaraMeshes(const GAME_OBJECT_ID obj_id)
-{
-    const bool lara_has_rifle = Inv_RequestItem(O_SHOTGUN_ITEM);
-    const bool lara_has_pistols = Inv_RequestItem(O_PISTOL_ITEM)
-        || Inv_RequestItem(O_MAGNUM_ITEM) || Inv_RequestItem(O_UZI_ITEM);
-
-    LARA_GUN_TYPE back_gun_type = LGT_UNARMED;
-    LARA_GUN_TYPE holsters_gun_type = LGT_UNARMED;
-
-    if (!lara_has_rifle && obj_id == O_SHOTGUN_ITEM) {
-        back_gun_type = LGT_SHOTGUN;
-    } else if (!lara_has_pistols && obj_id == O_PISTOL_ITEM) {
-        holsters_gun_type = LGT_PISTOLS;
-    } else if (!lara_has_pistols && obj_id == O_MAGNUM_ITEM) {
-        holsters_gun_type = LGT_MAGNUMS;
-    } else if (!lara_has_pistols && obj_id == O_UZI_ITEM) {
-        holsters_gun_type = LGT_UZIS;
-    }
-
-    if (back_gun_type != LGT_UNARMED) {
-        Gun_SetLaraBackMesh(back_gun_type);
-    }
-
-    if (holsters_gun_type != LGT_UNARMED) {
-        Gun_SetLaraHolsterLMesh(holsters_gun_type);
-        Gun_SetLaraHolsterRMesh(holsters_gun_type);
-    }
-}
