--- src/tr2/game/gun.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#include "game/gun/gun.h"
