--- src/tr2/game/gun/gun.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#include "global/types.h"
-
-#include <libtrx/game/gun.h>
-
-void Gun_UpdateLaraMeshes(GAME_OBJECT_ID obj_id);
