+++ src/tr2/game/gun/gun_rifle.c
@@ -1,12 +1,12 @@
 #include "game/gun/gun_rifle.h"
 
-#include "game/gun/gun.h"
 #include "game/gun/gun_misc.h"
 #include "game/lara/misc.h"
 #include "game/stats.h"
 
 #include <libtrx/config.h>
 #include <libtrx/game/game.h>
+#include <libtrx/game/gun.h>
 #include <libtrx/game/input.h>
 #include <libtrx/game/lara.h>
 #include <libtrx/game/lara/const.h>
