+++ tools/sort_imports
@@ -36,10 +36,6 @@ def main() -> None:
             "game/game/game_demo.c": "game/game.h",
             "game/game/game_draw.c": "game/game.h",
             "game/game/game_pause.c": "game/game.h",
-            "game/gun/gun.c": "game/gun.h",
-            "game/inventry.c": "game/inv.h",
-            "game/invfunc.c": "game/inv.h",
-            "game/invvars.c": "game/inv.h",
             "game/lara/lara.c": "game/lara.h",
             "game/option/option.c": "game/option.h",
             "game/savegame/savegame.c": "game/savegame.h",
